requirements.txt
pyqt6>=6.5.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
tflite-runtime>=2.14.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
